requirements.txt

torch

# -- Core Framework & Server --
Flask
gunicorn
waitress
Flask-Compress
# -- Background Tasks & Caching --
huey[redis]
redis

# -- YouTube & Data Scraping --
yt-dlp
youtube-transcript-api
beautifulsoup4
requests

# -- AI & Machine Learning --
--extra-index-url https://download.pytorch.org/whl/cpu
torch
numpy
openai
google-generativeai
langchain==0.0.354
sentence-transformers

# -- Google API (for YouTube Data) --
google-api-python-client
google-auth-httplib2
isodate

# -- Supabase & Database --
supabase==2.23.0
psycopg2-binary

# -- PDF Processing --
pymupdf

# -- Utilities --
python-dotenv
Markdown
MarkupSafe
PyJWT
cryptography


grpcio
isodate
httpx[http2]
tiktoken


discord
Flask-Mail
python-dateutil
razorpay
paypalrestsdk

# -- Production Server --
gevent
